Autocomplete (#107)
* feat(autocomplete): add endpoint

* feat(search): add throttle to search endpoint

* fix(search): remove unsused formet argument

* HTMX api call

* Remove HTMX and use Fetch API + UI

* search partial and display logic

* Fix post form

* reviews

* fix back

* use public_key from back

* hide source id when empty

* reviews 2

---------
