+++ assets/styles/index.css
@@ -403,7 +403,6 @@ Home Page
 
 .section-main {
     width: 100%;
-    overflow: hidden;
     position: relative;
 }
 
@@ -417,22 +416,25 @@ Home Page
 
 #hero-anim-map-container {
     position: absolute;
-    bottom: -38%;
-    right: 0;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    overflow: hidden;
+    pointer-events: none;
 }
 
 #hero-anim-map {
     transform: rotateZ(-23deg);
     position: absolute;
-    bottom: 0;
     right: -130px;
+    bottom: -18%;
     transition: all 0.4s ease-out;
     -webkit-box-shadow: -4px -1px 11px -2px rgba(0,0,0,0.25);
     box-shadow: -4px -1px 11px -2px rgba(0,0,0,0.25);
 }
 
 .section-main:hover #hero-anim-map {
-    bottom: 0;
     right: -100px;
     transform: rotateZ(-24deg);
 }
