+++ templates/header.html
@@ -29,11 +29,16 @@
                 <div class="fr-header__tools">
                     <div class="fr-header__tools-links">
                         <ul class="fr-btns-group">
-                            <li>
+                            {% if request.resolver_match.url_name not in 'home,create' %}
+                            <li class="d-none d-lg-block">
+                                {% include "project/partials/search.html" %}
+                            </li>
+                            <li class="d-lg-none d-xl-none">
                                 <a class="fr-btn" href="{% url 'project:create' %}">
                                     Créer un diagnostic
                                 </a>
                             </li>
+                        {% endif %}
                             {% if user.is_authenticated %}
                             <li>
                                 <button class="fr-btn" aria-expanded="false" aria-controls="menu-mon-compte">Mon compte</button>
